+++ CHANGELOG.rst
@@ -13,6 +13,9 @@ Added
 - DOI links to readme and documentation.
 - Area-based definition to ellipse geometry.
 - Flowchart example for standard workflow.
+- Unit tests.
+- ``position`` property to Ellipse class.
+- ``__eq__`` method to Ellipse class.
 
 Changed
 '''''''